Add more metadata to the db, adjust model.Recipe
This patch changes our database format to not contain the instructions
and adds some of the metadata from model.Recipe into it (namely:
timing and nutrition details).

The end result is a database file that's less than half the original
size:

933M    new/chronicle.db
2.0G    old/chronicle.db

+++ src/main/java/co/caio/cerberus/model/Recipe.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.util.List;
 import java.util.Map;
+import java.util.OptionalDouble;
 import java.util.OptionalInt;
 import org.immutables.value.Value;
 
@@ -36,11 +37,11 @@ public interface Recipe {
 
   OptionalInt calories();
 
-  OptionalInt carbohydrateContent();
+  OptionalDouble carbohydrateContent();
 
-  OptionalInt fatContent();
+  OptionalDouble fatContent();
 
-  OptionalInt proteinContent();
+  OptionalDouble proteinContent();
 
   @Value.Check
   default void check() {
